+++ app/page.tsx
@@ -1,4 +1,5 @@
 import Hero from "@/components/home/hero";
+import HowItWorks from "@/components/home/how-it-works";
 import ValueProp from "@/components/home/value-prop";
 
 export default function Home() {
@@ -10,6 +11,9 @@ export default function Home() {
 
         {/* Value Propositions Section */}
         <ValueProp />
+
+        {/* How it works */}
+        <HowItWorks />
       </main>
     </div>
   );